Share DOM-based bell icon and status-notice level type
Move the bell glyph and a stroke-icon builder into icons.ts, drop chatPanel's hard-coded bell HTML/innerHTML in favor of DOM construction, and reuse StatusNoticeLevel for the conversation bar badge.


Copilot-Session: 34fb7427-231a-4e48-b177-6db6750fd26d
